requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
streamlit
beautifulsoup4
faiss-cpu
langchain-ollama
